+++ Examples/MagicWeather/app/src/main/java/com/revenuecat/sample/extensions/Errors.kt
@@ -1,0 +1,12 @@
+package com.revenuecat.sample.extensions
+
+import android.app.AlertDialog
+import android.content.Context
+
+fun buildError(context: Context?, message: String) {
+    context?.let {
+        val alertDialogBuilder = AlertDialog.Builder(context)
+        alertDialogBuilder.setMessage(message)
+        alertDialogBuilder.show()
+    }
+}